- Top-level Package refactoring - Improved JWT Test - Switched to real-world public/private key of MIW
