+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/examples/BuildDIDDoc.java
@@ -22,10 +22,10 @@ package org.eclipse.tractusx.ssi.examples;
 import java.net.URI;
 import java.util.ArrayList;
 import java.util.List;
-import org.eclipse.tractusx.ssi.lib.base.MultibaseFactory;
 import org.eclipse.tractusx.ssi.lib.crypt.ed25519.Ed25519KeySet;
 import org.eclipse.tractusx.ssi.lib.did.web.DidWebFactory;
 import org.eclipse.tractusx.ssi.lib.model.MultibaseString;
+import org.eclipse.tractusx.ssi.lib.model.base.MultibaseFactory;
 import org.eclipse.tractusx.ssi.lib.model.did.Did;
 import org.eclipse.tractusx.ssi.lib.model.did.DidDocument;
 import org.eclipse.tractusx.ssi.lib.model.did.DidDocumentBuilder;
