+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/examples/ResolveDIDDoc.java
@@ -20,6 +20,7 @@
 package org.eclipse.tractusx.ssi.examples;
 
 import java.net.http.HttpClient;
+import org.eclipse.tractusx.ssi.lib.did.resolver.DidDocumentResolverRegistryImpl;
 import org.eclipse.tractusx.ssi.lib.did.web.DidWebDocumentResolver;
 import org.eclipse.tractusx.ssi.lib.did.web.DidWebFactory;
 import org.eclipse.tractusx.ssi.lib.did.web.util.DidWebParser;
@@ -27,7 +28,6 @@ import org.eclipse.tractusx.ssi.lib.exception.DidDocumentResolverNotRegisteredEx
 import org.eclipse.tractusx.ssi.lib.model.did.Did;
 import org.eclipse.tractusx.ssi.lib.model.did.DidDocument;
 import org.eclipse.tractusx.ssi.lib.model.did.DidMethod;
-import org.eclipse.tractusx.ssi.lib.resolver.DidDocumentResolverRegistryImpl;
 
 public class ResolveDIDDoc {
   public static DidDocument ResovleDocument(String didUrl)
