+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/lib/did/web/util/Ed25519PublicKeyParser.java
@@ -3,8 +3,8 @@ package org.eclipse.tractusx.ssi.lib.did.web.util;
 import java.io.StringReader;
 import lombok.SneakyThrows;
 import org.bouncycastle.util.io.pem.PemReader;
-import org.eclipse.tractusx.ssi.lib.base.MultibaseFactory;
 import org.eclipse.tractusx.ssi.lib.model.MultibaseString;
+import org.eclipse.tractusx.ssi.lib.model.base.MultibaseFactory;
 
 public class Ed25519PublicKeyParser {
