+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/lib/model/base/Base64.java
@@ -17,15 +17,38 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.ssi.lib.resolver;
+package org.eclipse.tractusx.ssi.lib.model.base;
 
-import org.eclipse.tractusx.ssi.lib.model.did.Did;
-import org.eclipse.tractusx.ssi.lib.model.did.DidDocument;
-import org.eclipse.tractusx.ssi.lib.model.did.DidMethod;
+import io.ipfs.multibase.Multibase;
+import lombok.EqualsAndHashCode;
+import lombok.NonNull;
+import lombok.Value;
+import org.eclipse.tractusx.ssi.lib.model.MultibaseString;
 
-public interface DidDocumentResolver {
+@Value
+@EqualsAndHashCode
+public class Base64 implements MultibaseString {
 
-  DidMethod getSupportedMethod();
+  public static boolean canDecode(String encoded) {
+    return Multibase.encoding(encoded).equals(Multibase.Base.Base64);
+  }
 
-  DidDocument resolve(Did did);
+  public static Base64 create(byte[] decoded) {
+    final String encoded = Multibase.encode(Multibase.Base.Base64, decoded);
+    return new Base64(decoded, encoded);
+  }
+
+  public static Base64 create(String encoded) {
+
+    if (!canDecode(encoded)) {
+      throw new IllegalArgumentException();
+    }
+
+    final byte[] decoded = Multibase.decode(encoded);
+
+    return new Base64(decoded, encoded);
+  }
+
+  byte @NonNull [] decoded;
+  @NonNull String encoded;
 }
