+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/lib/model/base/Base64WithPadding.java
@@ -17,10 +17,9 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-package org.eclipse.tractusx.ssi.lib.base;
+package org.eclipse.tractusx.ssi.lib.model.base;
 
 import io.ipfs.multibase.Multibase;
-import java.util.Objects;
 import lombok.EqualsAndHashCode;
 import lombok.NonNull;
 import lombok.Value;
@@ -28,30 +27,27 @@ import org.eclipse.tractusx.ssi.lib.model.MultibaseString;
 
 @Value
 @EqualsAndHashCode
-public class Base58Bitcoin implements MultibaseString {
+public class Base64WithPadding implements MultibaseString {
 
   public static boolean canDecode(String encoded) {
-    Objects.requireNonNull(encoded, "encoded must not be null");
-
-    return Multibase.encoding(encoded).equals(Multibase.Base.Base58BTC);
+    return Multibase.encoding(encoded).equals(Multibase.Base.Base64Pad);
   }
 
-  public static Base58Bitcoin create(byte[] decoded) {
-
-    final String encoded = Multibase.encode(Multibase.Base.Base58BTC, decoded);
-
-    return new Base58Bitcoin(decoded, encoded);
+  public static Base64WithPadding create(byte[] decoded) {
+    final String encoded = Multibase.encode(Multibase.Base.Base64Pad, decoded);
+    return new Base64WithPadding(decoded, encoded);
   }
 
-  public static Base58Bitcoin create(String encoded) {
+  public static Base64WithPadding create(String encoded) {
 
     if (!canDecode(encoded)) {
-      throw new IllegalArgumentException("Encoded base58 String not in Base58BTC format");
+      throw new IllegalArgumentException(
+          "Encoded base64 String not in Base64 format (with padding)");
     }
 
-    final byte[] decoded = Multibase.decode(encoded);
+    final byte[] base64 = Multibase.decode(encoded);
 
-    return new Base58Bitcoin(decoded, encoded);
+    return new Base64WithPadding(base64, encoded);
   }
 
   byte @NonNull [] decoded;
