+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/lib/model/did/DidDocument.java
@@ -23,7 +23,7 @@ import java.net.URI;
 import java.util.*;
 import lombok.ToString;
 import org.eclipse.tractusx.ssi.lib.model.JsonLdObject;
-import org.eclipse.tractusx.ssi.lib.util.SerializeUtil;
+import org.eclipse.tractusx.ssi.lib.serialization.SerializeUtil;
 
 // spec https://www.w3.org/TR/did-core/
 @ToString
