+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/lib/proof/LinkedDataProofGenerator.java
@@ -22,10 +22,10 @@ package org.eclipse.tractusx.ssi.lib.proof;
 import java.net.URI;
 import java.time.Instant;
 import lombok.RequiredArgsConstructor;
-import org.eclipse.tractusx.ssi.lib.base.MultibaseFactory;
 import org.eclipse.tractusx.ssi.lib.model.Ed25519Signature2020;
 import org.eclipse.tractusx.ssi.lib.model.Ed25519Signature2020Builder;
 import org.eclipse.tractusx.ssi.lib.model.MultibaseString;
+import org.eclipse.tractusx.ssi.lib.model.base.MultibaseFactory;
 import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredential;
 import org.eclipse.tractusx.ssi.lib.proof.hash.HashedLinkedData;
 import org.eclipse.tractusx.ssi.lib.proof.hash.LinkedDataHasher;
