+++ cx-ssi-lib/src/main/java/org/eclipse/tractusx/ssi/lib/serialization/jwt/DanubTechVerifiableCredentialSerializerImpl.java
@@ -21,7 +21,7 @@ package org.eclipse.tractusx.ssi.lib.serialization.jwt;
 
 import java.util.Map;
 import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredential;
-import org.eclipse.tractusx.ssi.lib.serialization.jsonLd.DanubTechMapper;
+import org.eclipse.tractusx.ssi.lib.serialization.jsonLd.DanubeTechMapper;
 
 public class DanubTechVerifiableCredentialSerializerImpl implements VerifiableCredentialSerializer {
   @Override
@@ -30,14 +30,14 @@ public class DanubTechVerifiableCredentialSerializerImpl implements VerifiableCr
     final com.danubetech.verifiablecredentials.VerifiableCredential dtCredential =
         com.danubetech.verifiablecredentials.VerifiableCredential.fromMap(credentialJson);
 
-    return DanubTechMapper.map(dtCredential);
+    return DanubeTechMapper.map(dtCredential);
   }
 
   @Override
   public String serialize(VerifiableCredential credential) {
 
     final com.danubetech.verifiablecredentials.VerifiableCredential dtCredential =
-        DanubTechMapper.map(credential);
+        DanubeTechMapper.map(credential);
 
     return dtCredential.toJson(true);
   }
