+++ cx-ssi-lib/src/test/java/org/eclipse/tractusx/ssi/lib/serialization/jwt/SerializedJwtPresentationFactoryImplTest.java
@@ -1,42 +1,36 @@
 package org.eclipse.tractusx.ssi.lib.serialization.jwt;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import com.nimbusds.jwt.SignedJWT;
 import java.net.URI;
-import java.time.Instant;
 import java.util.List;
-import java.util.Map;
 import lombok.SneakyThrows;
 import org.eclipse.tractusx.ssi.lib.SsiLibrary;
 import org.eclipse.tractusx.ssi.lib.crypt.ed25519.Ed25519Key;
+import org.eclipse.tractusx.ssi.lib.did.resolver.OctetKeyPairFactory;
 import org.eclipse.tractusx.ssi.lib.jwt.SignedJwtFactory;
+import org.eclipse.tractusx.ssi.lib.jwt.SignedJwtVerifier;
 import org.eclipse.tractusx.ssi.lib.model.Ed25519Signature2020;
 import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredential;
-import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredentialBuilder;
-import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredentialSubject;
-import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredentialType;
 import org.eclipse.tractusx.ssi.lib.proof.LinkedDataProofGenerator;
-import org.eclipse.tractusx.ssi.lib.proof.LinkedDataProofValidation;
 import org.eclipse.tractusx.ssi.lib.proof.hash.LinkedDataHasher;
 import org.eclipse.tractusx.ssi.lib.proof.transform.LinkedDataTransformer;
 import org.eclipse.tractusx.ssi.lib.proof.verify.LinkedDataSigner;
-import org.eclipse.tractusx.ssi.lib.proof.verify.LinkedDataVerifier;
-import org.eclipse.tractusx.ssi.lib.resolver.OctetKeyPairFactory;
 import org.eclipse.tractusx.ssi.lib.serialization.jsonLd.JsonLdSerializerImpl;
 import org.eclipse.tractusx.ssi.lib.util.identity.TestDidDocumentResolver;
 import org.eclipse.tractusx.ssi.lib.util.identity.TestIdentity;
 import org.eclipse.tractusx.ssi.lib.util.identity.TestIdentityFactory;
+import org.eclipse.tractusx.ssi.lib.util.vc.TestCredentialFactory;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 class SerializedJwtPresentationFactoryImplTest {
 
-  private LinkedDataProofValidation linkedDataProofValidation;
   private LinkedDataProofGenerator linkedDataProofGenerator;
 
   private TestIdentity credentialIssuer;
   private TestDidDocumentResolver didDocumentResolver;
+
+  private SignedJwtVerifier jwtVerifier;
 
   @SneakyThrows
   @Test
@@ -46,12 +40,7 @@ class SerializedJwtPresentationFactoryImplTest {
 
     credentialIssuer = TestIdentityFactory.newIdentityWithED25519Keys(false);
     didDocumentResolver.register(credentialIssuer);
-
-    linkedDataProofValidation =
-        new LinkedDataProofValidation(
-            new LinkedDataHasher(),
-            new LinkedDataTransformer(),
-            new LinkedDataVerifier(didDocumentResolver.withRegistry()));
+    jwtVerifier = new SignedJwtVerifier(didDocumentResolver.withRegistry());
     linkedDataProofGenerator =
         new LinkedDataProofGenerator(
             new LinkedDataHasher(), new LinkedDataTransformer(), new LinkedDataSigner());
@@ -61,13 +50,15 @@ class SerializedJwtPresentationFactoryImplTest {
         credentialIssuer.getDidDocument().getVerificationMethods().get(0).getId();
     final byte[] privateKey = credentialIssuer.getPrivateKey();
 
-    final VerifiableCredential credential = createCredential(null);
+    final VerifiableCredential credential =
+        TestCredentialFactory.createCredential(credentialIssuer, null);
 
     final Ed25519Signature2020 proof =
         linkedDataProofGenerator.createEd25519Signature2020(
             credential, verificationMethod, privateKey);
 
-    final VerifiableCredential credentialWithProof = createCredential(proof);
+    final VerifiableCredential credentialWithProof =
+        TestCredentialFactory.createCredential(credentialIssuer, proof);
 
     SerializedJwtPresentationFactory presentationFactory =
         new SerializedJwtPresentationFactoryImpl(
@@ -82,26 +73,8 @@ class SerializedJwtPresentationFactoryImplTest {
             List.of(credentialWithProof),
             "test-audience",
             Ed25519Key.asPrivateKey(credentialIssuer.getPrivateKey()));
+
     Assertions.assertNotNull(presentation);
-    // @TODO: add JWT Signature Checks and Structural Validation
-  }
-
-  @SneakyThrows
-  private VerifiableCredential createCredential(Ed25519Signature2020 proof) {
-    final VerifiableCredentialBuilder verifiableCredentialBuilder =
-        new VerifiableCredentialBuilder();
-
-    final VerifiableCredentialSubject verifiableCredentialSubject =
-        new VerifiableCredentialSubject(Map.of("foo", "bar"));
-
-    return verifiableCredentialBuilder
-        .id(URI.create("did:test:id"))
-        .type(List.of(VerifiableCredentialType.VERIFIABLE_CREDENTIAL))
-        .issuer(credentialIssuer.getDid().toUri())
-        .expirationDate(Instant.parse("2025-02-15T17:21:42Z").plusSeconds(3600))
-        .issuanceDate(Instant.parse("2023-02-15T17:21:42Z"))
-        .proof(proof)
-        .credentialSubject(verifiableCredentialSubject)
-        .build();
+    Assertions.assertDoesNotThrow(() -> jwtVerifier.verify(presentation));
   }
 }
