+++ cx-ssi-lib/src/test/java/org/eclipse/tractusx/ssi/lib/util/vc/TestCredentialFactory.java
@@ -1,0 +1,36 @@
+package org.eclipse.tractusx.ssi.lib.util.vc;
+
+import java.net.URI;
+import java.time.Instant;
+import java.util.List;
+import java.util.Map;
+import lombok.SneakyThrows;
+import org.eclipse.tractusx.ssi.lib.model.Ed25519Signature2020;
+import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredential;
+import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredentialBuilder;
+import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredentialSubject;
+import org.eclipse.tractusx.ssi.lib.model.verifiable.credential.VerifiableCredentialType;
+import org.eclipse.tractusx.ssi.lib.util.identity.TestIdentity;
+
+public class TestCredentialFactory {
+
+  @SneakyThrows
+  public static VerifiableCredential createCredential(
+      TestIdentity issuer, Ed25519Signature2020 proof) {
+    final VerifiableCredentialBuilder verifiableCredentialBuilder =
+        new VerifiableCredentialBuilder();
+
+    final VerifiableCredentialSubject verifiableCredentialSubject =
+        new VerifiableCredentialSubject(Map.of("foo", "bar"));
+
+    return verifiableCredentialBuilder
+        .id(URI.create("did:test:id"))
+        .type(List.of(VerifiableCredentialType.VERIFIABLE_CREDENTIAL))
+        .issuer(issuer.getDid().toUri())
+        .expirationDate(Instant.parse("2025-02-15T17:21:42Z").plusSeconds(3600))
+        .issuanceDate(Instant.parse("2023-02-15T17:21:42Z"))
+        .proof(proof)
+        .credentialSubject(verifiableCredentialSubject)
+        .build();
+  }
+}
