+++ cx-ssi-lib/src/test/resources/keys/ed25519/private.pem
@@ -1,3 +1,4 @@
 -----BEGIN PRIVATE KEY-----
-MC4CAQAwBQYDK2VwBCIEIFgZFKsyHZ0yA9xkroZPB8NTIl1dopnX9nhH2q6puEK8
+MFECAQEwBQYDK2VwBCIEIGOZOZFEO6s1YpstydhDGzAGfHZKn1hXDFyz0g60Rqkv
+gSEAxU+mIr4sHAbqmAVndmjE+sCT80fgR1vb912l5FhRQWs=
 -----END PRIVATE KEY-----
