+++ cx-ssi-lib/src/test/resources/keys/ed25519/public.pem
@@ -1,3 +1,3 @@
 -----BEGIN PUBLIC KEY-----
-MCowBQYDK2VwAyEABqAmUe/amV/nAVUt01XyrLpmQLOyLqF6LnAkH4QdyqI=
+MCowBQYDK2VwAyEAxU+mIr4sHAbqmAVndmjE+sCT80fgR1vb912l5FhRQWs=
 -----END PUBLIC KEY-----
